+++ frontend/src/contexts/AuthContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, Dispatch, ReactNode, SetStateAction, useEffect, useState } from "react";
+import { createContext, Dispatch, SetStateAction } from "react";
 
 interface AuthContextProp{
   isAuth: boolean|null;
@@ -11,16 +11,3 @@ export const AuthContext = createContext<AuthContextProp>({
 });
 
 
-interface AuthProviderProp {
-  children: ReactNode;
-}
-export default function AuthProvider({ children }: AuthProviderProp) {
-  const [isAuth, setIsAuth] = useState<boolean | null>(null);
-
-  useEffect(() => {
-    const token = localStorage.getItem("token");
-    setIsAuth(!!token);
-  }, []);
-
-  return <AuthContext.Provider value={{ isAuth, setIsAuth }}>{children}</AuthContext.Provider>;
-}
