+++ frontend/src/routes/pages/Map3dPage.tsx
@@ -28,9 +28,10 @@ export default function Map3dPage() {
   const navigate = useNavigate();
 
   useEffect(()=>{
-    if(!isAuth){
+    const token = localStorage.getItem("token")
+    if(!token){
       alert("Signing in is required");
-      navigate("/")
+      navigate("/");
     }
   },[isAuth, navigate])
 
@@ -48,7 +49,7 @@ export default function Map3dPage() {
     enabled: !!selectedOperationAndDate,
   });
 
-  if (isPending) return "Loading...";
+  // if (isPending) return "Loading...";
   if (error) {
     if (error.message === "Unauthorized user") {
       localStorage.removeItem("token");
