+++ frontend/src/routes/pages/MapPage.tsx
@@ -29,7 +29,8 @@ export default function MapPage() {
 
 
   useEffect(()=>{
-    if(!isAuth){
+    const token = localStorage.getItem("token")
+    if(!token){
       alert("Signing in is required");
       navigate("/");
     }
@@ -49,7 +50,7 @@ export default function MapPage() {
     enabled: !!selectedOperationAndDate,
   });
 
-  if (isPending) return "Loading...";
+  // if (isPending) return "Loading...";
   if (error) {
     if (error.message === "Unauthorized user") {
       localStorage.removeItem("token");
