+++ frontend/src/types/latLonAlt.ts
@@ -2,7 +2,7 @@ export interface LatLonAlt {
   lat: number; //위도
   lon: number; //경도
   alt: number; //고도
-  heading: number | null;
+  // heading: number | null;
 }
 export interface LatLon {
   lat: number; //위도
